requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
audioread==3.0.1
    # via librosa
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via soundfile
charset-normalizer==3.4.2
    # via requests
decorator==5.2.1
    # via librosa
googleapis-common-protos==1.70.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.71.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via ollama
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
joblib==1.5.0
    # via
    #   librosa
    #   scikit-learn
lazy-loader==0.4
    # via librosa
librosa==0.11.0
    # via lunar_tools_art (pyproject.toml)
llvmlite==0.44.0
    # via numba
msgpack==1.1.1
    # via librosa
numba==0.61.2
    # via librosa
numpy==2.2.5
    # via
    #   librosa
    #   numba
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
ollama==0.5.1
    # via lunar_tools_art (pyproject.toml)
opentelemetry-api==1.34.1
    # via
    #   lunar_tools_art (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.34.1
    # via lunar_tools_art (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.34.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.34.1
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.34.1
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.55b1
    # via opentelemetry-instrumentation-logging
opentelemetry-instrumentation-logging==0.55b1
    # via lunar_tools_art (pyproject.toml)
opentelemetry-proto==1.34.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.34.1
    # via
    #   lunar_tools_art (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.55b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==24.2
    # via
    #   lazy-loader
    #   opentelemetry-instrumentation
    #   pooch
platformdirs==4.3.8
    # via pooch
pooch==1.8.2
    # via librosa
protobuf==5.29.5
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pycparser==2.22
    # via cffi
pydantic==2.11.4
    # via ollama
pydantic-core==2.33.2
    # via pydantic
pydub==0.25.1
    # via lunar_tools_art (pyproject.toml)
python-dotenv==1.1.0
    # via lunar_tools_art (pyproject.toml)
requests==2.32.3
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   pooch
scikit-learn==1.7.0
    # via librosa
scipy==1.15.3
    # via
    #   librosa
    #   scikit-learn
sniffio==1.3.1
    # via anyio
soundfile==0.13.1
    # via librosa
soxr==0.5.0.post1
    # via librosa
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.13.2
    # via
    #   anyio
    #   librosa
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
wrapt==1.17.2
    # via opentelemetry-instrumentation
zipp==3.21.0
    # via importlib-metadata
